Module jf_aap::proof

SNARK proofs data structures, generation and verification API

How to use proof-related API

use crate::proof;
use crate::proof::transfer::{self, TransferWitness, TransferPublicInput};
use crate::utils::compute_universal_param_size,
use crate::structs::NoteType;

let rng = &mut ark_std::test_rng();

let num_input = 3;
let num_output = 3;
let depth = 30;
let max_degree = compute_universal_param_size(NoteType::Transfer, num_input, num_output, depth).unwrap();
// Step 1. Get universal parameters for all types of transaction
let universal_param = proof::universal_setup(max_degree, rng).unwrap();

// Step 2. Preprocess for specific transaction type (e.g. Transfer)
let (proving_key, verifying_key) =
  transfer::preprocess(rng, &universal_param, num_input, num_output, depth).unwrap();

// Step 3. Prover generate valid proof
let witness = TransferWitness::new_unchecked(...);
let public_inputs = TransferPublicInput::from_witness(&witness, ...);
let validity_proof = transfer::prove(rng, &proving_key, &witness, &public_inputs).unwrap();

// Step 4. Verifier verifies the proof
assert_eq!(transfer::verify(&verifying_key, &public_inputs, &validity_proof).unwrap(), true);

Modules

freeze

This module contains functions and data structures for

mint

This module contains functions and data structures for

transfer

This module contains functions and data structures for

Functions

load_srs

Use Common Reference String parameters from Aztec’s MPC ceremony in proving validity of all transactions, regardless of the transaction type.

universal_setup

One-time universal setup for parameters to be used in proving validity of all transactions, regardless of the transaction type.

Type Definitions

UniversalParam

The universal parameters (Structured Reference String) for proving/verifying transaction validity of any type. Generate only once during a multi-party setup. For testing purpose, use universal_setup() for a one-time generation.