Module jf_aap::structs

record-related data structures in transactions

Structs

AssetCode

Asset code structure

AssetCodeSeed

The random seed used in AssetCode derivation

AssetDefinition

Asset Definition

AssetPolicy

Policies related to asset record

AuditData

Transfer AuditMemo decrypted

AuditMemo

Memos for auditors such as auditors required by the asset policy. Concretely, it is a ciphertext over details of a transaction, enabling asset tracing and identity tracing.

BlindFactor

The blind factor used to produce a hiding commitment

ExpirableCredential

A credential with expiry issued by a credential issuer for a user testifying user’s identity attributes

FeeInput

All necessary information for the input record that is meant to pay transaction fee in all different transactions.

IdentityAttribute

An identity attribute of a user, usually attested via ExpirableCredential issued by an identity issuer.

InternalAssetCode

A unique identifier/code for an asset type

Nullifier

The nullifier represents a spent/consumed asset record

ReceiverMemo

The proof of membership in an accumulator (Merkle tree) for an asset record Encrypted Message for owners of transaction outputs

RecordCommitment

Asset record to be published

RecordOpening

The opening of an asset record, containing all fields (and secrets) required to compute commitment/proofs/.. values related to this asset record.

RevealMap

A bitmap indicating which of the following fields are to be revealed: (upk_x, upk_y, v, blind, attrs) where reveal bits for upk_x and upk_y are the same. Also note that asset code code is compulsorily revealed.

TxnFeeInfo

Fee structure containing fee input spending info, fee to pay and change record opening

Enums

FreezeFlag

Flag indicating whether records is frozen or not

NoteType

Enum for each type of note